Populate apputils with functions that are not routes
